Finish delete functionality
Delete a character is a really funny thing! It's easy to delete a
character which is not at the head of line. We just use `memmove` to
reallocate the memory, and move the cursor to the right one step.

The trivial thing is when we delete the character at the first line, we
can delete `\n`, so we need to first free the memory this row used. And
put the original remained characters to the up line.

What a wonderful thing!