Finish syntax highlighting
At this point, the whole kilo tutorial is finished. The most important
thing of an editor is just to store where the cursor is. This is the
basic functionality.

And next we need to make the memory allocation, how to store the string
in the buffer, so we first should construct a array which contains the
row pointer and the row size.

For each row, we need to store the string, it is easy all. And we also
should use a render string to just render the original string, what a
wonderful idea.

And for syntax highlighting, we should use a syntax array to map with
the render string.

Well, it's a wonderful time to code an editor in the terminal.